backend/requirements.txt
fastapi
uvicorn
ultralytics
Pillow
torch
